refactor: extract pkg/k8s/scanner
