+++ pkg/k8s/scanner/io.go
@@ -1,4 +1,4 @@
-package k8s
+package scanner
 
 import (
 	"fmt"
